% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontdesk.R
\name{frontdesk}
\alias{frontdesk}
\title{Frontdesk Shiny Application}
\usage{
frontdesk(clientName = Sys.getenv("CannaData_clientName"),
  host = Sys.getenv("CannaData_host"),
  port = as.integer(Sys.getenv("CannaData_port")),
  user = Sys.getenv("CannaData_user"),
  password = Sys.getenv("CannaData_password"),
  base_url = Sys.getenv("CannaData_baseUrl"),
  db = Sys.getenv("CannaData_db"),
  bucket = Sys.getenv("CannaData_AWS_bucket"),
  budtender = Sys.getenv("SHINYPROXY_USERNAME"))
}
\arguments{
\item{clientName}{Name of dispensary}

\item{host}{MySQL enpoint}

\item{port}{MySQL port}

\item{user}{MySQL user}

\item{password}{MySQL password}

\item{base_url}{Where Shiny Server is hosted}

\item{db}{MySQL db}

\item{bucket}{Name of AWS bucket}

\item{budtender}{Name of user}
}
\description{
Frontdesk Shiny Application
}
