% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontdesk.R
\name{frontdesk}
\alias{frontdesk}
\title{Frontdesk Shiny Application}
\usage{
frontdesk(clientName = getOption("CannaData_clientName"),
  host = getOption("CannaData_host"),
  port = as.integer(getOption("CannaData_port")),
  user = getOption("CannaData_user"),
  password = getOption("CannaData_password"),
  base_url = getOption("CannaData_baseUrl"), db = getOption("CannaData_db"),
  bucket = getOption("CannaData_AWS_bucket"),
  budtender = getOption("SHINYPROXY_USERNAME"))
}
\arguments{
\item{clientName}{Name of dispensary}

\item{host}{MySQL enpoint}

\item{port}{MySQL port}

\item{user}{MySQL user}

\item{password}{MySQL password}

\item{base_url}{Where Shiny Server is hosted}

\item{db}{MySQL db}

\item{bucket}{Name of AWS bucket}

\item{budtender}{Name of user}
}
\description{
Frontdesk Shiny Application
}
