% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontdesk.R
\name{frontdesk}
\alias{frontdesk}
\title{Frontdesk Shiny Application}
\usage{
frontdesk(pool = pool::dbPool(RMariaDB::MariaDB(), host =
  getOption("CannaData_host"), port = as.integer(getOption("CannaData_port")),
  user = getOption("CannaData_user"), password =
  getOption("CannaData_password"), db = getOption("CannaData_db")),
  clientName = getOption("CannaData_clientName"),
  base_url = getOption("CannaData_baseUrl"),
  bucket = getOption("CannaData_AWS_bucket"),
  auth_id = getOption("auth_id"), auth_secret = getOption("auth_secret"),
  scope = "openid email",
  connection_name = "Username-Password-Authentication")
}
\arguments{
\item{pool}{pool connection}

\item{clientName}{Name of dispensary}

\item{base_url}{Where Shiny Server is hosted}

\item{bucket}{Name of AWS bucket}

\item{auth_id}{Auth0 ID}

\item{auth_secret}{Auth0 Secret}

\item{scope}{scope}

\item{connection_name}{connect name}
}
\description{
Frontdesk Shiny Application
}
