% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontdesk.R
\name{frontdesk}
\alias{frontdesk}
\title{Frontdesk Shiny Application}
\usage{
frontdesk(pool = pool::dbPool(RMariaDB::MariaDB(), host =
  getOption("CannaData_host"), port = as.integer(getOption("CannaData_port")),
  user = getOption("CannaData_user"), password =
  getOption("CannaData_password"), db = getOption("CannaData_db"), `ssl-ca` =
  "https://s3.amazonaws.com/rds-downloads/rds-combined-ca-bundle.pem",
  `ssl-verify` = "server-cert"),
  clientName = getOption("CannaData_clientName"),
  base_url = getOption("CannaData_baseUrl"),
  bucket = getOption("CannaData_AWS_bucket"),
  gcp_json = getOption("CannaData_google_json"),
  auth_id = getOption("auth_id"), auth_secret = getOption("auth_secret"),
  scope = "openid email", twilio_sid = getOption("TWILIO_SID"),
  twilio_token = getOption("TWILIO_TOKEN"),
  connection_name = getOption("auth_connection"),
  state = getOption("CannaData_state"))
}
\arguments{
\item{pool}{pool connection}

\item{clientName}{Name of dispensary}

\item{base_url}{Where Shiny Server is hosted}

\item{bucket}{Name of AWS bucket}

\item{gcp_json}{Google Cloud Print Json}

\item{auth_id}{Auth0 ID}

\item{auth_secret}{Auth0 Secret}

\item{scope}{scope}

\item{twilio_sid}{Twilio secret id}

\item{twilio_token}{Twilio token}

\item{connection_name}{connect name}
}
\description{
Frontdesk Shiny Application
}
